fix: UserService Interface 분리
