+++ src/main/java/com/study/todoparty/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.study.todoparty.controller;
 
-import com.study.todoparty.service.UserService;
+import com.study.todoparty.service.UserServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,5 +9,5 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RequestMapping("/users")
 public class UserController {
-    private final UserService userService;
+    private final UserServiceImpl userServiceImpl;
 }
